refactor of UI
