+++ src/swop/Main/Main.java
@@ -1,6 +1,6 @@
 package swop.Main;
 
-import swop.UI.TempUI;
+import swop.UI.LoginUI;
 
 /**
  * Main class to run the project
@@ -13,6 +13,6 @@ public class Main {
     public static void main(String[] args) {
         AssemAssist assemAssist = new AssemAssist();
 //        assemAssist.run();
-        new TempUI(assemAssist);
+        new LoginUI(assemAssist);
     }
 }
