+++ src/swop/Tests/UseCaseTests/CheckProductionStatisticsTest.java
@@ -3,19 +3,14 @@ package swop.Tests.UseCaseTests;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import java.io.ByteArrayInputStream;
-import java.io.ByteArrayOutputStream;
-import java.io.PrintStream;
 import java.util.Arrays;
 import java.util.ListIterator;
 
 import org.junit.jupiter.api.Test;
 
 import swop.Main.AssemAssist;
-import swop.UI.LoginUI;
 import swop.UI.Builders.DisplayStatus;
 import swop.UI.Generators.ManagerGenerator;
-import swop.UI.TempManUI;
-import swop.UI.TempUI;
 import swop.Users.Manager;
 
 public class CheckProductionStatisticsTest {
