+++ src/swop/Tests/UseCaseTests/OrderNewCarTest.java
@@ -8,17 +8,11 @@ import swop.Car.CarModel.ModelA;
 import swop.Car.CarModel.ModelB;
 import swop.Car.CarModel.ModelC;
 import swop.Main.AssemAssist;
-import swop.UI.LoginUI;
 import swop.UI.Builders.DisplayStatus;
 import swop.UI.Generators.GarageHolderGenerator;
-import swop.UI.TempUI;
-import swop.Users.CarMechanic;
 import swop.Users.GarageHolder;
 
-import java.io.ByteArrayInputStream;
-import java.io.ByteArrayOutputStream;
 import java.io.InputStream;
-import java.io.PrintStream;
 import java.util.Arrays;
 import java.util.List;
 import java.util.ListIterator;
